refactor : login.js 모듈화
tags.js : 공통으로 사용되는 태그들
addEvents.js : 공통으로 사용되는 태그에 addEventListener
function.js : 공통으로 사용되는 함수들
signin.js : 로그인에만 필요한 로직
signup.js : 회원가입에만 필요한 로직
