+++ pages/login/signin.html
@@ -49,6 +49,7 @@
             </div>
         </main>
     </div>
-    <script src="login.js"></script>
+    <script type="module" src="addEvents.js"></script>
+    <script type="module" src="signin.js"></script>
 </body>
 </html>
