+++ pages/login/signup.html
@@ -54,6 +54,7 @@
             </div>
         </main>
     </div>
-    <script src="login.js"></script>
+    <script type="module" src="addEvents.js"></script>
+    <script type="module" src="signup.js"></script>
 </body>
 </html>
