+++ pages/login/tags.js
@@ -1,0 +1,4 @@
+const form = document.querySelector('#form');
+const hidePasswordButton = document.querySelector('.hide-password');
+
+export { form, hidePasswordButton };
